moved changed to Edgescan pack
