+++ Packs/Edgescan/ReleaseNotes/1_0_3.md
@@ -1,0 +1,4 @@
+
+#### Integrations
+##### Edgescan
+- Added the **edgescan-vulnerabilities-add-annotation** command for adding annotations into vulnerabilities.
