--- Packs/EdgescanV2/Integrations/Edgescan/Pipfile
@@ -1,23 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-pylint = "*"
-pytest = "*"
-pytest-mock = "*"
-requests-mock = "*"
-pytest-asyncio = "*"
-pytest-xdist = "*"
-pytest-datadir-ng = "*"
-freezegun = "*"
-pytest-json = "*"
-vcrpy = "*"
-pytest-cov = "*"
-hypothesis = "*"
-
-[packages]
-
-[requires]
-python_version = "3.9"
