--- Packs/EdgescanV2/pack_metadata.json
@@ -1,17 +1,0 @@
-{
-    "name": "Edgescan",
-    "description": "Cloud-based continuous vulnerability management and penetration testing solution.",
-    "support": "community",
-    "currentVersion": "1.0.0",
-    "author": "Maciej Drobniuch",
-    "url": "",
-    "email": "",
-    "created": "2021-08-09T09:40:13Z",
-    "categories": [],
-    "tags": [],
-    "useCases": [],
-    "keywords": [],
-    "githubUser": [
-        "mdrobniu"
-    ]
-}
